LayoutTests/imported/w3c/web-platform-tests/feature-policy/reporting/microphone-reporting.https-expected.txt

CONSOLE MESSAGE: ReferenceError: Can't find variable: setMediaPermission

Harness Error (FAIL), message = ReferenceError: Can't find variable: setMediaPermission

NOTRUN Microphone Report Format
